+++ lib/legion/api.rb
@@ -54,6 +54,7 @@ require_relative 'api/library_routes'
 require_relative 'api/sync_dispatch'
 require_relative 'api/lex_dispatch'
 require_relative 'api/tbi_patterns'
+require_relative 'api/inbound_webhooks'
 require_relative 'api/graphql' if defined?(GraphQL)
 
 module Legion
@@ -180,6 +181,7 @@ module Legion
     register Routes::Knowledge
     register Routes::Logs
     register Routes::TbiPatterns
+    register Routes::InboundWebhooks
     register Routes::GraphQL if defined?(Routes::GraphQL)
 
     use Legion::API::Middleware::RequestLogger